Add privacy policy accepted time field in Users object
