+++ src/main/java/uk/ac/cam/cl/dtg/isaac/dto/users/UserSummaryForAdminUsersDTO.java
@@ -23,7 +23,7 @@ import java.time.Instant;
  */
 public class UserSummaryForAdminUsersDTO extends UserSummaryWithEmailAddressDTO {
   private Instant lastUpdated;
-  private Instant privacyPolicyAcceptedTime;
+//  private Instant privacyPolicyAcceptedTime;
   private Instant lastSeen;
   private Instant registrationDate;
   private String schoolId;
@@ -54,23 +54,23 @@ public class UserSummaryForAdminUsersDTO extends UserSummaryWithEmailAddressDTO 
     this.lastUpdated = lastUpdated;
   }
 
-  /**
-   * Gets the privacyPolicyAcceptedTime.
-   *
-   * @return the privacyPolicyAcceptedTime
-   */
-  public Instant getPrivacyPolicyAcceptedTime() {
-    return privacyPolicyAcceptedTime;
-  }
-
-  /**
-   * Sets the privacyPolicyAcceptedTime.
-   *
-   * @param privacyPolicyAcceptedTime the privacyPolicyAcceptedTime to set
-   */
-  public void setPrivacyPolicyAcceptedTime(final Instant privacyPolicyAcceptedTime) {
-    this.privacyPolicyAcceptedTime = privacyPolicyAcceptedTime;
-  }
+//  /**
+//   * Gets the privacyPolicyAcceptedTime.
+//   *
+//   * @return the privacyPolicyAcceptedTime
+//   */
+//  public Instant getPrivacyPolicyAcceptedTime() {
+//    return privacyPolicyAcceptedTime;
+//  }
+//
+//  /**
+//   * Sets the privacyPolicyAcceptedTime.
+//   *
+//   * @param privacyPolicyAcceptedTime the privacyPolicyAcceptedTime to set
+//   */
+//  public void setPrivacyPolicyAcceptedTime(final Instant privacyPolicyAcceptedTime) {
+//    this.privacyPolicyAcceptedTime = privacyPolicyAcceptedTime;
+//  }
 
   /**
    * Gets the lastSeen.
@@ -148,7 +148,7 @@ public class UserSummaryForAdminUsersDTO extends UserSummaryWithEmailAddressDTO 
   public String toString() {
     return "lastUpdated=" + "UserSummaryForAdminUsersDTO{" + lastUpdated
         + ", lastSeen=" + lastSeen
-        + ", privacyPolicyAcceptedTime=" + privacyPolicyAcceptedTime
+//        + ", privacyPolicyAcceptedTime=" + privacyPolicyAcceptedTime
         + ", registrationDate=" + registrationDate
         + ", schoolId=" + schoolId
         + ", schoolOther=" + schoolOther
